backend/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
djangorestframework-simplejwt
argon2-cffi
requests
python-dotenv
channels[daphne]
channels
drf-yasg
